% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/href_goalie-stats.R
\name{href_goalie_stats}
\alias{href_goalie_stats}
\title{Scrape NHL Goalie Stats from hockey-reference.com}
\usage{
href_goalie_stats(seas, reg = TRUE, post = TRUE)
}
\arguments{
\item{seas}{Which seasons would you like to retrieve data for? No default set.}

\item{reg}{Include regular season data? Defaults to TRUE}

\item{post}{Include postseason data? Defaults to TRUE}
}
\description{
Gathers NHL goalie stats by season for the selected seasons and combines results into one dataset.
}
\examples{
href_goalie_stats()
}
\keyword{skillalytics}
