% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/href_schedule-results.R
\name{href_results}
\alias{href_results}
\title{Scrape seasonal NHL game results from hockey-reference.com}
\usage{
href_results(seas, reg = TRUE, post = TRUE)
}
\arguments{
\item{seas}{Which seasons would you like to retrieve data for? No default set.}

\item{reg}{Include regular season data? Defaults to TRUE}

\item{post}{Include postseason data? Defaults to TRUE}
}
\description{
Gathers year-by-year game results from each NHL season selected and combines results into one dataset.
}
\examples{
href_results()
}
\keyword{skillalytics}
