% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/href_standings.R
\name{href_standings}
\alias{href_standings}
\title{Scrape NHL Standings from hockey-reference.com}
\usage{
href_standings(seas)
}
\arguments{
\item{seas}{Which seasons would you like to retrieve data for? No default set.}
}
\description{
Gathers year-by-year standings from each NHL season selected and combines results into one dataset.
}
\examples{
href_standings()
}
\keyword{skillalytics}
