% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{skillalytics_setup}
\alias{skillalytics_setup}
\title{Setup function for running all other Skillalytics R functions}
\usage{
skillalytics_setup(seas)
}
\arguments{
\item{seas}{Which seasons would you like to retrieve data for? No default set.}
}
\description{
This ensures proper setup of seasons selected to scrape and team name/ids for scraping
and organizing datasets.
}
\examples{
skillalytics_setup()
}
\keyword{skillalytics}
